Move orders storage into orders folder
